refactor(AA-30): add interfaces to service layer
